backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart
pydantic>=2.0.0
openai>=1.0.0
oss2>=2.18.0
pydicom>=2.4.0
pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
reportlab>=4.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
